Backend/src/main/java/com/example/E_Commerce/service/impl/AddressServiceImpl.java
package com.example.E_Commerce.service.impl;

import com.example.E_Commerce.dto.AddressDto;
import com.example.E_Commerce.dto.Response;
import com.example.E_Commerce.entity.Address;
import com.example.E_Commerce.entity.User;
import com.example.E_Commerce.repository.AddressRepo;
import com.example.E_Commerce.service.AddressService;
import com.example.E_Commerce.service.UserService;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class AddressServiceImpl implements AddressService {

    private final AddressRepo addressRepo;
    private final UserService userService;


    @Override
    public Response saveAndUpdateAddress(AddressDto addressDto) {
        User user = userService.getLoginUser();
        Address address = user.getAddress();

        if (address == null) {
            address = new Address();
            address.setUser(user);
        }

        if (addressDto.getStreet() != null) address.setStreet(addressDto.getStreet());
        if (addressDto.getCity() != null) address.setCity(addressDto.getCity());
        if (addressDto.getState() != null) address.setState(addressDto.getState());
        if (addressDto.getZipCode() != null) address.setZipCode(addressDto.getZipCode());
        if (addressDto.getCountry() != null) address.setCountry(addressDto.getCountry());

        addressRepo.save(address);

        String message = user.getAddress() == null ?
                "Address successfully created" :
                "Address successfully updated";

        return Response.builder()
                .status(200)
                .message(message)
                .build();
    }
}
